move two another headers to src/utils
