+++ src/fmstream.cpp
@@ -22,8 +22,8 @@
 
 #include "fmstream.h"
 
-#include "align.h"
 #include "exception_control/string_exception.h"
+#include "utils/align.h"
 #include "utils/value_size_defines.h"
 
 #include <algorithm>
