+++ src/hdstream.cpp
@@ -22,9 +22,9 @@
 
 #include "hdstream.h"
 
-#include "align.h"
 #include "id3v2tag.h"
 #include "exception_control/string_exception.h"
+#include "utils/align.h"
 #include "utils/value_size_defines.h"
 
 #include <algorithm>
