+++ src/hdstream.h
@@ -28,7 +28,7 @@
 #include "props.h"
 #include "pvrstream.h"
 #include "rtldevice.h"
-#include "scalar_condition.h"
+#include "utils/scalar_condition.h"
 
 #include <atomic>
 #include <condition_variable>
