+++ src/tcpdevice.cpp
@@ -36,9 +36,9 @@
 #include <sys/socket.h>
 #include <unistd.h>
 #endif
-#include "align.h"
 #include "exception_control/socket_exception.h"
 #include "exception_control/string_exception.h"
+#include "utils/align.h"
 
 #include <cstring>
