+++ src/utils/CMakeLists.txt
@@ -1,5 +1,7 @@
 set(SOURCES complex.cpp)
 
-set(HEADERS value_size_defines.h)
+set(HEADERS align.h
+            scalar_condition.h
+            value_size_defines.h)
 
 add_library(code_src_utils OBJECT ${SOURCES} ${HEADERS})
