+++ src/utils/scalar_condition.h
