+++ src/wxstream.h
@@ -29,7 +29,7 @@
 #include "props.h"
 #include "pvrstream.h"
 #include "rtldevice.h"
-#include "scalar_condition.h"
+#include "utils/scalar_condition.h"
 
 #include <atomic>
 #include <condition_variable>
